+++ app/Modules/TimeTracking/Views/time_tracking/project_create.blade.php
@@ -21,6 +21,8 @@
             @lang('bt.create_project')
         </h3>
         <div class="float-right">
+            <a class="btn btn-warning float-right" href={!! route('timeTracking.projects.index')  !!}><i
+                        class="fa fa-ban"></i> @lang('bt.cancel')</a>
             <button class="btn btn-primary"><i class="fa fa-save"></i> @lang('bt.save')</button>
         </div>
         <div class="clearfix"></div>